requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5

motor==3.6.0
pymongo==4.9.1

py7zr==0.21.0
rarfile==4.2
PyPDF2==3.0.1
m3u8==5.0.0

aiohttp==3.10.11
psutil==6.1.0
python-dotenv==1.0.1

fastapi==0.115.0
uvicorn==0.30.6

yt-dlp>=2024.1.1
Pillow>=10.0.0
mutagen>=1.47.0
